More secret sharing refactoring. #11
