+++ include/nil/crypto3/pubkey/secret_sharing_state.hpp
